+++ sinp_nomenclatures/views.py
@@ -6,14 +6,13 @@ import logging
 
 from django.db.models import Prefetch
 from django_filters.rest_framework import DjangoFilterBackend
-
-# from django.contrib.auth.mixins import LoginRequiredMixin
-from rest_framework.permissions import IsAuthenticated
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
 from .models import Nomenclature, Source, Type
+from .permissions import IsOptionnalyAuthenticated
 from .serializers import (
     NomenclatureSerializer,
+    NomenclatureWithParentsSerializer,
     SourceSerializer,
     TypeSerializer,
     TypeSerializerWithNomenclature,
@@ -26,18 +25,33 @@ class NomenclatureViewset(ReadOnlyModelViewSet):
     """Nomenclature item viewset (ReadOnly)"""
 
     serializer_class = NomenclatureSerializer
-    permission_classes = [IsAuthenticated]
-    queryset = Nomenclature.objects.all()
+    permission_classes = [IsOptionnalyAuthenticated]
+    queryset = Nomenclature.objects.all().prefetch_related("parents")
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ["type", "code", "status", "parents"]
+
+    def get_serializer_class(self):
+        """Choose simple serializer or serializer serializer with parents"""
+        with_parents = self.request.query_params.get(
+            "with_parents", default=False
+        )
+        return (
+            NomenclatureWithParentsSerializer
+            if with_parents
+            else NomenclatureSerializer
+        )
 
 
 class TypeViewset(ReadOnlyModelViewSet):
     """Nomenclature type viewset (ReadOnly)"""
 
     serializer_class = TypeSerializer
-    permission_classes = [IsAuthenticated]
-    queryset = Type.objects.all()
+    permission_classes = [IsOptionnalyAuthenticated]
+    queryset = (
+        Type.objects.all()
+        .prefetch_related("nomenclatures")
+        .prefetch_related("nomenclatures__parents")
+    )
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ["code", "mnemonic", "source", "status"]
 
@@ -77,7 +91,7 @@ class SourceViewset(ReadOnlyModelViewSet):
     """Source viewset (ReadOnly)"""
 
     serializer_class = SourceSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsOptionnalyAuthenticated]
     queryset = Source.objects.all()
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ["name"]